backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-bcrypt
flask-cors
python-dotenv
pymysql
faker
pytest